+++ src/app/app.component.html
@@ -1,1 +1,3 @@
+ 
+<h1>New Structure</h1>
 <router-outlet></router-outlet>
