+++ src/app/backoffice/pages/activities/activity-form/activity-form.component.ts
@@ -7,10 +7,18 @@ import { Component, OnInit } from '@angular/core';
 })
 export class ActivityFormComponent implements OnInit {
   title = 'base-ong-angular-client';
-
+  uploadedFiles: any[] = [];
   constructor() { }
 
   ngOnInit(): void {
   }
+
+  onUpload(event:any) {
+        for(let file of event.files) {
+            this.uploadedFiles.push(file);
+        }
+
+        // this.messageService.add({severity: 'info', summary: 'File Uploaded', detail: ''});
+    }
 
 }
