+++ src/app/backoffice/pages/auth/register-form/register-form.component.html
@@ -1,0 +1,1 @@
+<p>register-form works!</p>
