--- src/app/core/core.module.ts
@@ -1,13 +1,0 @@
-import { CommonModule } from "@angular/common";
-import { HttpClientModule } from "@angular/common/http";
-import {
-  NgModule
-} from "@angular/core";
-import { HttpService } from "./services/http.service";
-
-@NgModule({
-  declarations: [],
-  imports: [CommonModule, HttpClientModule],
-  providers: [HttpService],
-})
-export class CoreModule {}
