--- src/app/core/services/http.service.spec.ts
@@ -1,16 +1,0 @@
-import { TestBed } from '@angular/core/testing';
-
-import { HttpService } from './http.service';
-
-describe('HttpService', () => {
-  let service: HttpService;
-
-  beforeEach(() => {
-    TestBed.configureTestingModule({});
-    service = TestBed.inject(HttpService);
-  });
-
-  it('should be created', () => {
-    expect(service).toBeTruthy();
-  });
-});
