--- src/app/features/app-routing.module.ts
@@ -1,33 +1,0 @@
-import { ActivityFormComponent } from "./pages/activities/activity-form/activity-form.component";
-import { NgModule } from "@angular/core";
-import { CommonModule } from "@angular/common";
-import { RouterModule, Routes } from "@angular/router";
-import { RegisterFormComponent } from "./pages/auth/register-form/register-form.component";
-
-const routes: Routes = [
-  {
-    path: "actividades",
-    component: ActivityFormComponent
-  },
-  {
-    path: "registro",
-    component: RegisterFormComponent,
-    pathMatch: "full",
-  },
-  {
-    path: "",
-    redirectTo: "actividades",
-    pathMatch: "full",
-  },
-  {
-    path: "**",
-    redirectTo: "actividades",
-    pathMatch: "full",
-  },
-];
-
-@NgModule({
-  declarations: [],
-  imports: [CommonModule, RouterModule.forRoot(routes)],
-})
-export class AppRoutingModule {}
