--- src/app/features/features.module.ts
@@ -1,48 +1,0 @@
-import { RouterModule } from "@angular/router";
-import { CommonModule } from "@angular/common";
-import { NgModule } from "@angular/core";
-import { AppRoutingModule } from "./app-routing.module";
-import { ActivityFormComponent } from "./pages/activities/activity-form/activity-form.component";
-import { LoginFormComponent } from "./pages/auth/login-form/login-form.component";
-import { RegisterFormComponent } from "./pages/auth/register-form/register-form.component";
-import { CategoriesFormComponent } from "./pages/categories/categories-form/categories-form.component";
-import { NewsFormComponent } from "./pages/news/news-form/news-form.component";
-import { SlidesFormComponent } from "./pages/slides/slides-form/slides-form.component";
-import { TestimonialFormComponent } from "./pages/testimonials/testimonial-form/testimonial-form.component";
-import { UserFormComponent } from "./pages/users/user-form/user-form.component";
-import { InputTextModule } from 'primeng/inputtext';
-import { ButtonModule } from 'primeng/button';
-import { ReactiveFormsModule } from "@angular/forms";
-
-@NgModule({
-  declarations: [
-    ActivityFormComponent,
-    LoginFormComponent,
-    RegisterFormComponent,
-    CategoriesFormComponent,
-    NewsFormComponent,
-    SlidesFormComponent,
-    TestimonialFormComponent,
-    UserFormComponent,
-  ],
-  exports: [
-    ActivityFormComponent,
-    LoginFormComponent,
-    RegisterFormComponent,
-    CategoriesFormComponent,
-    NewsFormComponent,
-    SlidesFormComponent,
-    TestimonialFormComponent,
-    UserFormComponent,
-    RouterModule
-  ],
-  imports: [
-    CommonModule,
-    AppRoutingModule,
-    RouterModule,
-    ReactiveFormsModule,
-    InputTextModule,
-    ButtonModule
-  ]
-})
-export class FeaturesModule {}
