--- src/app/features/pages/activities/activity-form/activity-form.component.html
@@ -1,2 +1,0 @@
-<p>hola mundo</p>
-<h1>Hola</h1>
